requirements.txt
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==2.8
oauthlib==3.1.0
protobuf==4.21.9
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==3.0.9
PyYAML==6.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
uritemplate==4.1.1
urllib3==1.25.6
